+++ denotational/core/src/test/kotlin/io/github/oxidefrp/core/test_framework/shared/SubjectSpec.kt
@@ -4,11 +4,11 @@ import io.github.oxidefrp.core.Cell
 import io.github.oxidefrp.core.EventStream
 import io.github.oxidefrp.core.Moment
 import io.github.oxidefrp.core.cutOff
-import io.github.oxidefrp.core.test_framework.CellValidator
-import io.github.oxidefrp.core.test_framework.EventStreamValidator
-import io.github.oxidefrp.core.test_framework.MomentValidator
 import io.github.oxidefrp.core.test_framework.Validator
-import io.github.oxidefrp.core.test_framework.ValueValidator
+import io.github.oxidefrp.core.test_framework.validators.CellValidator
+import io.github.oxidefrp.core.test_framework.validators.EventStreamValidator
+import io.github.oxidefrp.core.test_framework.validators.MomentValidator
+import io.github.oxidefrp.core.test_framework.validators.ValueValidator
 
 /**
  * A specification of a test subject, i.e. a set of expectations what the
